renamed logosection classes as "logos" and completed logos classes
