+++ configuration.properties
@@ -5,7 +5,7 @@ env = QA
 cookiesEnableDisable = 2
 #cookies enable =2 , disable =0
 
-browserClose = 1
+browserClose = 0
 #0 means close the browser after test
 
 #smokeTestTime = 12   #Smoke test can run every 6 hours
