+++ src/test/java/proqa/pages/ServicesPage.java
@@ -1,0 +1,4 @@
+package proqa.pages;
+
+public class ServicesPage {
+}
