--- src/test/resources/features/LogoSection.feature
@@ -1,20 +1,0 @@
-Feature: Verifying Logo Section
-
-  Scenario Outline:Logo Section Verification
-
-    Given User is on Landing page
-    When Verifies that "<Logos>" are present
-    Examples:
-      | Logos       |
-      |Selenium     |
-      |BitBucket    |
-      |cucumber     |
-      |Jira Software|
-      |Jenkins      |
-      |appium       |
-      |Maven        |
-      |Azure        |
-      |Java         |
-      |Rest-assured |
-      |Apache jMeter|
-      |TestNG       |
