+++ src/test/resources/features/Logos.feature
@@ -1,0 +1,7 @@
+Feature:Verifying Logo Section
+
+  @logos
+  Scenario:Logo Section Verification
+
+    Given User is on Landing page
+    When Verifies that Logos are present
