+++ src/test/resources/features/Services.feature
@@ -1,0 +1,3 @@
+Feature: Verifying Services Elements
+
+  Scenario: Services elements verification
